refactor: update UI file structure
Declutter root folder by moving UI-related settings to respective folder.
Group UI files together based on specific page.
