--- .stylelintignore
@@ -1,9 +1,0 @@
-.history
-.husky
-.vscode
-coverage
-dist
-node_modules
-vite.config.ts
-jest.config.ts
-src-tauri
