+++ package.json
@@ -6,17 +6,16 @@
   "scripts": {
     "start": "tauri dev",
     "build": "tauri build",
-    "start:front": "vite",
-    "build:front": "tsc && vite build",
-    "preview:front": "vite preview",
-    "lint:scripts": "eslint . --ext .ts",
-    "lint:styles": "stylelint ./**/*.less",
+    "start:front": "vite ./src-ui",
+    "build:front": "tsc && vite build ./src-ui",
+    "lint:scripts": "eslint ./src-ui --ext .ts",
+    "lint:styles": "stylelint ./src-ui/**/*.less",
     "lint:rust": "cd src-tauri && cargo clippy",
     "lint": "npm run lint:scripts && npm run lint:styles && npm run lint:rust",
     "format:scripts": "prettier . --write --list-different",
     "format:scripts:check": "prettier . -c",
-    "format:styles": "stylelint ./**/*.less --fix",
-    "format:styles:check": "stylelint ./**/*.less",
+    "format:styles": "stylelint ./src-ui/**/*.less --fix",
+    "format:styles:check": "stylelint ./src-ui/**/*.less",
     "format:rust": "cd src-tauri && cargo fmt --all",
     "format:rust:check": "cd src-tauri && cargo fmt --all -- --check",
     "format": "npm run format:rust && npm run format:scripts && npm run format:styles",
