+++ src-ui/.stylelintignore
@@ -1,0 +1,2 @@
+coverage
+dist
