+++ src-ui/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="./src/styles/main.less" />
+    <link rel="stylesheet" href="./index.less" />
   </head>
   <body>
     <div id="app">
@@ -71,6 +71,6 @@
         />
       </svg>
     </div>
-    <script type="module" src="/src/main.ts"></script>
+    <script type="module" src="./index.ts"></script>
   </body>
 </html>
