+++ src-ui/settings/settings.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="src/styles/settings.less" />
+    <link rel="stylesheet" href="./settings.less" />
   </head>
   <body>
     <form id="settings-form">
@@ -47,6 +47,6 @@
         /></label>
       </div>
     </form>
-    <script type="module" src="/src/settings.ts"></script>
+    <script type="module" src="./settings.ts"></script>
   </body>
 </html>
