+++ src-ui/settings/settings.less
@@ -1,5 +1,5 @@
-@import './theme.less';
-@import './common.less';
+@import '../styles/theme.less';
+@import '../styles/common.less';
 
 html {
   background: var(--background);
