+++ src-ui/vite.config.ts
@@ -1,4 +1,5 @@
 import { resolve } from 'path';
+
 import { defineConfig } from 'vite';
 
 export default defineConfig({
@@ -23,7 +24,7 @@ export default defineConfig({
     rollupOptions: {
       input: {
         main: resolve(__dirname, 'index.html'),
-        settings: resolve(__dirname, 'settings.html'),
+        settings: resolve(__dirname, 'settings/settings.html'),
       },
     },
   },
